pyproject.toml
[tool.poetry]
name = "Burhan"
version = "0.1.0"
description = "Islamic QA System - Multi-Agent Architecture based on Fanar-Sadiq"
authors = ["Burhan Team <team@Burhan.islamic>"]
readme = "README.md"
packages = [{include = "src", from = "."}]
include = ["tests/"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
redis = {extras = ["hiredis"], version = "^5.2.0"}
qdrant-client = "^1.12.0"
openai = "^1.58.0"
groq = "^0.13.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
structlog = "^24.4.0"
pyyaml = "^6.0.2"
psycopg2-binary = "^2.9.11"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.7.0"
mypy = "^1.13.0"

[tool.poetry.group.rag]
optional = true

[tool.poetry.group.rag.dependencies]
torch = ">=2.2.0"
transformers = ">=4.40.0"
accelerate = ">=0.34.0"

[tool.poetry.group.data]
optional = true

[tool.poetry.group.data.dependencies]
ijson = "^3.2.3"

[tool.poetry.group.eval]
optional = true

[tool.poetry.group.eval.dependencies]
ragas = "^0.1.7"
datasets = "^2.19.0"
langchain = "^0.1.0"
langchain-openai = "^0.1.0"

[tool.poetry.scripts]
Burhan = "src.api.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = ["asyncpg", "redis", "qdrant_client"]
ignore_missing_imports = true
